+++ here-naksha-lib-psql/src/commonTest/kotlin/naksha/psql/UpdateFeatureTest.kt
@@ -2,7 +2,6 @@ package naksha.psql
 
 import naksha.base.Action
 import naksha.base.Guid
-import naksha.base.Int64
 import naksha.base.NakshaError
 import naksha.base.TupleNumber
 import naksha.base.Version
@@ -130,9 +129,9 @@ class UpdateFeatureTest : PgTestBase(collection = null, catalogId = "") {
         assertEquals(createdTuple.getByteArray(naksha.model.objects.XyzMembers.XyzReferencePoint), updatedTuple.getByteArray(naksha.model.objects.XyzMembers.XyzReferencePoint))
         assertNull(createdTuple.decodeFeature(null)?.properties["new_attr"])
         assertEquals("some_value", updatedTuple.decodeFeature(null)?.properties["new_attr"])
-        assertEquals(createdTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt)?.let { if (it == Int64(0L)) null else it } ?: createdTuple.getLong(naksha.model.objects.XyzMembers.XyzUpdatedAt), updatedTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt)?.let { if (it == Int64(0L)) null else it })
+        assertEquals(createdTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt)?.let { if (it == 0L) null else it } ?: createdTuple.getLong(naksha.model.objects.XyzMembers.XyzUpdatedAt), updatedTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt)?.let { if (it == 0L) null else it })
         assertNotEquals(updatedTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt), updatedTuple.getLong(naksha.model.objects.XyzMembers.XyzUpdatedAt))
-        assertNull(createdTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt)?.let { if (it == Int64(0L)) null else it })
+        assertNull(createdTuple.getLong(naksha.model.objects.XyzMembers.XyzCreatedAt)?.let { if (it == 0L) null else it })
         assertNotNull(createdTuple.getLong(naksha.model.objects.XyzMembers.XyzUpdatedAt))
         assertEquals(createdTuple.getInt(naksha.model.objects.XyzMembers.XyzHereTile), updatedTuple.getInt(naksha.model.objects.XyzMembers.XyzHereTile))
         assertEquals(Action.UPDATE, updatedTuple.tupleNumber.action)
@@ -159,7 +158,7 @@ class UpdateFeatureTest : PgTestBase(collection = null, catalogId = "") {
         testWithCollection("atomicUpdateNotExistingWithFakeUuid")
 
         val featureId = "feature_not_existing"
-        val fakeUUID = TupleNumber(storage.number, catalog.catalogNumber, collection.collectionNumber, Naksha.featureNumber(featureId), Version.now(Int64(1), Action.CREATE).number)
+        val fakeUUID = TupleNumber(storage.number, catalog.catalogNumber, collection.collectionNumber, Naksha.featureNumber(featureId), Version.now(1L, Action.CREATE).number)
         val feature = NakshaFeature().apply {
             id = featureId
             properties.xyz.setRaw("uuid", fakeUUID)
